Rename routes, deprecate old post route
